move filter related files to the folder
