+++ src/components/views/projects/filterContainer/ProjectsFiltersMobile.tsx
@@ -1,9 +1,9 @@
 import ProjectsSearchTablet from '@/components/views/projects/ProjectsSearchTablet';
 import { Flex } from '@/components/styled-components/Flex';
 import ProjectsSubCategories from '@/components/views/projects/ProjectsSubCategories';
-import ProjectsFiltersSwiper from '@/components/views/projects/ProjectsFiltersSwiper';
+import ProjectsFiltersSwiper from '@/components/views/projects/filterContainer/ProjectsFiltersSwiper';
 import { StyledLine } from '@/components/views/projects/common.styled';
-import ProjectsFiltersButton from '@/components/views/projects/ProjectsFiltersButton';
+import ProjectsFiltersButton from '@/components/views/projects/filterContainer/ProjectsFiltersButton';
 
 const ProjectsFiltersMobile = () => {
 	return (
